# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto API
  description: |
    This is the API for the WASAphoto Homework 1 for the WaSa course by 1849661.
  version: 0.0.1
...
paths:
...
  /users/{id}/banned_users/{banned_user_id}:
    parameters:
        - $ref: "#/components/parameters/banned_user_id"
        - $ref: '#/components/parameters/identifier'
  
    put:
      tags: ["ban"]
      summary: Bans a user
      description: Once a user is blocked he/she won't be able to see anything from whoever banned him/her. That includes comments,likes, photos and the profile of the banning user. Banning multiple times is equivalent to banning a user once
      operationId: banUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
    
    delete:
      tags: ["ban"]
      summary: Unbans a user
      description: By removing the ban the unbanned user can see the posts,comments and likes of the user that unbanned him/her
      operationId: unbanUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
#=====================================================================================
  /users/{id}/followers/{follower_id}:
    parameters:
        - $ref: "#/components/parameters/follower_id"
        - $ref: '#/components/parameters/identifier'
        
    put:
      tags: ["followers"]
      summary: Follows a certain user
      description: |
        Allows a user (if he's logged in) to follow another user only if the latter didn't ban him. A user can't follow himself and can't follow a user he already follows. Once a user follows another one then his list of "Following" will be updated too
      operationId: followUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
          
    delete:
      tags: ["followers"]
      summary: Unfollows a certain user
      description: |
        A user can decide to unfollow another one. He will be able to follow him again in the future
      operationId: unfollowUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"      
          
      security:
        - bearerAuth: [] 
#=====================================================================================
...
  /users/{id}/photos/{photo_id}/likes/{like_id}:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        - $ref: '#/components/parameters/like_id'
  
    put:
      tags: ["likes"]
      summary: Add a like to a photo
      description: Adds a like to a user's photo. A user can only like a photo once (multiple likes count as one). Owners can't like their own photos
      operationId: likePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
    
    delete:
      tags: ["likes"]
      summary: Remove a like from a photo
      description: Removes a like from a user's photo
      operationId: unlikePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
#=====================================================================================
#_____________________________________________________________________________________________________
components:
  parameters:
    identifier:
      name: id
      in: path
      description: A user's unique id
      required: true
      schema:
        description: String between 3-16 characters
        type: string
        pattern: '^.*?$'
        example: Marietto64
        minLength: 3
        maxLength: 16
        readOnly: true
#........................................................        
    photo_id:
      name: photo_id
      in: path
      description: A photo unique identifier
      required: true
      schema: 
        description: Int64 between 0-9999
        type: integer
        format: int64
        minimum: 0
        maximum: 9999
        example : 999
        readOnly: true
#........................................................      
...
    follower_id:
      name: follower_id
      in: path
      description: A follower's unique identifier. It's exactly his *identifier*
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Piccioni"
#........................................................       
    like_id:
      name: like_id
      in: path
      description: A like unique identifier. It's exactly the *identifier* of the user that liked the photo
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "LauraZ"
#........................................................       
    banned_user_id:
      name: banned_user_id
      in: path
      description: A banned user unique identifier. It's exactly the *identifier* of the user
      required: true
      schema:
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Akille"
#........................................................      
...
  responses:
...
    no_content:
      description: Response associated to the 204 http status (No content to send for this reques)
#''''''''''''''''''''''''''''''''''''''''''''''''''''''''     
...
    unauthorized:
      description:  Response associated to the 401 http status (Access token is missing. User is not logged in)
#''''''''''''''''''''''''''''''''''''''''''''''''''''''''          
    forbidden:
      description: Response associated to the 403 http status (Access denied )
#''''''''''''''''''''''''''''''''''''''''''''''''''''''''     
    not_found:
      description: Response associated to the 404 http status (The requested resource doesn't exist)
#''''''''''''''''''''''''''''''''''''''''''''''''''''''''            
    internal_server_error:
      description: Response associated to the 500 http status (Server has encountered an unknown error)
#''''''''''''''''''''''''''''''''''''''''''''''''''''''''
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
#_____________________________________________________________________________________________________